Move EventManager to gaphor.core
