+++ gaphor/UML/tests/test_modelfactory.py
@@ -1,8 +1,8 @@
 import pytest
 
 from gaphor import UML
+from gaphor.core.eventmanager import EventManager
 from gaphor.core.modeling import ElementFactory
-from gaphor.services.eventmanager import EventManager
 
 
 @pytest.fixture
