+++ gaphor/storage/tests/test_verify.py
@@ -1,6 +1,6 @@
 from gaphor import UML
+from gaphor.core.eventmanager import EventManager
 from gaphor.core.modeling import ElementFactory
-from gaphor.services.eventmanager import EventManager
 from gaphor.storage.verify import orphan_references
