+++ gaphor/ui/tests/test_elementeditor.py
@@ -1,7 +1,7 @@
 import pytest
 
+from gaphor.core.eventmanager import EventManager
 from gaphor.core.modeling import ElementFactory
-from gaphor.services.eventmanager import EventManager
 from gaphor.ui.elementeditor import ElementEditor
 from gaphor.ui.mainwindow import Diagrams
