+++ pyproject.toml
@@ -67,7 +67,7 @@ gaphorconvert = 'gaphor.tools.gaphorconvert:main'
 
 [tool.poetry.plugins."gaphor.services"]
 "component_registry" = "gaphor.services.componentregistry:ComponentRegistry"
-"event_manager" = "gaphor.services.eventmanager:EventManager"
+"event_manager" = "gaphor.core.eventmanager:EventManager"
 "properties" = "gaphor.services.properties:Properties"
 "undo_manager" = "gaphor.services.undomanager:UndoManager"
 "element_factory" = "gaphor.core.modeling:ElementFactory"
